+++ sdk/iot_hub/src/service/operations/apply_on_edge_device.rs
