+++ sdk/iot_hub/src/service/operations/create_or_update_configuration.rs
