+++ sdk/iot_hub/src/service/operations/delete_configuration.rs
