+++ sdk/iot_hub/src/service/operations/delete_identity.rs
