+++ sdk/iot_hub/src/service/operations/get_identity.rs
