+++ sdk/iot_hub/src/service/operations/query.rs
